+++ src/plonemeeting/portal/core/testing.py
@@ -33,8 +33,13 @@ class PlonemeetingPortalCoreLayer(PloneSandboxLayer):
 
         utils.getRequest = patched_getRequest
 
+        # Run the startup hook
+        from plone.app.theming.plugins.hooks import onStartup
+        onStartup(None)
+
     def setUpPloneSite(self, portal):
         applyProfile(portal, "plonemeeting.portal.core:default")
+        applyProfile(portal, 'plone.app.theming:default')
         transaction.commit()