+++ src/plonemeeting/portal/core/migrations/migrate_to_1004.py
@@ -43,13 +43,17 @@ class MigrateTo1004(Migrator):
         logger.info("Migrating to plonemeeting.portal 1004...")
         self._init_long_representatives_in_charge()
         self._apply_folder_constraints()
+        self.refreshDatabase(catalogs=False,
+                             workflows=True,
+                             workflowsToUpdate=["institution_workflow"])
 
 
 def migrate(context):
     """
     This migration function will:
-       1) Initialize the new attribute long_representatives_in_charge ;
-       2) Apply contraints on Folders ;
+       1) Initialize the new attribute long_representatives_in_charge;
+       2) Apply contraints on Folders;
+       3) Update security for elements using institution_workflow.
     """
     migrator = MigrateTo1004(context)
     migrator.run()